Fix issue with custom functions compiled from expression
When a function is compiled from an expression during runtime
the .NET framework adds a "hidden" parameter to the function
of type ClosureInfo. This is an internal compiler detail but
it is exposed when fetching the parameters from the MethodInfo.
This causes Jace to think there is an extra parameter when the
compiled expression is used as a custom function.